Debug Admin

+++ commandes/reload.js
@@ -7,6 +7,11 @@ module.exports = {
   },
 
   run: async function (client, message, args) {
+    if (
+      !message.member.hasPermission('BAN_MEMBERS') &&
+      message.author.id !== '297388490595762186'
+    )
+      return message.channel.send("tu n'as pas la permission de faire ça !");
     await message.delete();
     await client.channels.cache
       .get('749855316543537163')
